Revert "gettin down on some input loop shenanigans"
This reverts commit 587dc56cb0f6a20bc93c6dd54627a150811a2d91.

+++ src/ui/panel.rs
@@ -12,7 +12,7 @@ pub struct Panel {
     pub border: CharRenderer,
     pub background: CharRenderer,
     pub widgets: Vec<Widget>,
-    pub id: i32,
+    pub id: u64,
 }
 
 impl Panel {
@@ -22,7 +22,7 @@ impl Panel {
         bounds: Vector2,
         border: CharRenderer,
         background: CharRenderer,
-        id: i32
+        id: u64
     ) -> Self {
         let title = String::from(title);
         Self {
